move WeightWithDynamicFloat8CastTensor to fsdp_utils.py
Summary:

Refactor in preparation of adding delayed scaling support to float8
all-gather, no logic change.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

[ghstack-poisoned]
